+++ Cargo.toml
@@ -33,7 +33,6 @@ libp2p = { default-features = false, features = ["floodsub", "identify", "kad", 
 nom = "6"
 opentelemetry = { version = "0.17.0", features = ["rt-tokio"] }
 opentelemetry-jaeger = { version = "0.16.0", features = ["rt-tokio", "reqwest_collector_client"] }
-# opentelemetry-otlp = { version = "0.10.0", features = ["rt-tokio"] }
 prometheus = { version = "0.13.0", features = ["process"] }
 regex = "1.5"
 reqwest = { version = "0.11", features = ["json"] }
@@ -53,7 +52,6 @@ tracing-log = "0.1"
 tracing-opentelemetry = "0.17.2"
 tracing-subscriber = { version = "0.3.11", features = ["env-filter", "json"] }
 urlencoding = "2.1"
-uuid = { version = "1.0.0", features = ["v4"] }
 void = "1"
 
 [dev-dependencies]
