+++ src/lib.rs
@@ -29,7 +29,6 @@ mod tracing;
 pub mod transport;
 pub mod zcap;
 
-use ::tracing::subscriber::set_global_default;
 use libp2p::{
     identity::{ed25519::Keypair as Ed25519Keypair, Keypair},
     PeerId,
@@ -37,41 +36,14 @@ use libp2p::{
 use relay::RelayNode;
 use routes::{cors, delegate, invoke, open_host_key, relay_addr};
 use std::{collections::HashMap, sync::RwLock};
-use tracing_log::LogTracer;
-use tracing_subscriber::{layer::SubscriberExt, Layer, Registry};
 
 #[get("/healthz")]
 pub fn healthcheck() {}
 
-pub fn tracing_try_init(config: &config::Logging) {
-    LogTracer::init().unwrap();
-    let env_filter = tracing_subscriber::EnvFilter::from_default_env();
-    let subscriber = tracing_subscriber::fmt::layer();
-    let log = match config.format {
-        config::LoggingFormat::Text => subscriber.boxed(),
-        config::LoggingFormat::Json => subscriber.json().boxed(),
-    };
-    let telemetry = if config.tracing.enabled {
-        let tracer = opentelemetry_jaeger::new_pipeline()
-            .with_service_name("kepler")
-            .install_batch(opentelemetry::runtime::Tokio)
-            .unwrap();
-        let telemetry = tracing_opentelemetry::layer().with_tracer(tracer);
-        Some(telemetry)
-    } else {
-        None
-    };
-    let collector = Registry::default()
-        .with(env_filter)
-        .with(log)
-        .with(telemetry);
-    set_global_default(collector).unwrap();
-}
-
 pub async fn app(config: &Figment) -> Result<Rocket<Build>> {
     let kepler_config = config.extract::<config::Config>()?;
 
-    tracing_try_init(&kepler_config.log);
+    tracing::tracing_try_init(&kepler_config.log);
 
     storage::KV::healthcheck(kepler_config.storage.indexes.clone()).await?;
     storage::StorageUtils::new(kepler_config.storage.blocks.clone())
